+++ tests/test_validators.py
@@ -3,33 +3,14 @@ from wsgiref.validate import validator
 
 import pytest
 from sigma.exceptions import SigmaValueError
-from sigma.modifiers import (
-    SigmaAllModifier,
-    SigmaBase64OffsetModifier,
-    SigmaContainsModifier,
-)
+
 from sigma.rule import SigmaDetectionItem, SigmaLogSource, SigmaRule
 from sigma.types import SigmaString
+
 from sigma.validators.core.logsources import (
     SpecificInsteadOfGenericLogsourceValidator,
     SpecificInsteadOfGenericLogsourceIssue,
 )
-
-from sigma.validators.core.condition import (
-    AllOfThemConditionIssue,
-    AllOfThemConditionValidator,
-    DanglingDetectionIssue,
-    DanglingDetectionValidator,
-    ThemConditionWithSingleDetectionIssue,
-    ThemConditionWithSingleDetectionValidator,
-)
-from sigma.validators.core.modifiers import (
-    AllWithoutContainsModifierIssue,
-    Base64OffsetWithoutContainsModifierIssue,
-    InvalidModifierCombinationsValidator,
-    ModifierAppliedMultipleIssue,
-)
-
 from sigma.validators.core.values import (
     ControlCharacterIssue,
     ControlCharacterValidator,
@@ -97,147 +78,6 @@ def rules_with_id_collision():
         )
         for i in range(2)
     ]
-
-
-def test_validator_dangling_detection():
-    validator = DanglingDetectionValidator()
-    rule = SigmaRule.from_yaml(
-        """
-    title: Test
-    status: test
-    logsource:
-        category: test
-    detection:
-        referenced1:
-            field1: val1
-        referenced2:
-            field2: val2
-        referenced3:
-            field3: val3
-        unreferenced:
-            field4: val4
-        condition: (referenced1 or referenced2) and referenced3
-    """
-    )
-    assert validator.validate(rule) == [DanglingDetectionIssue([rule], "unreferenced")]
-
-
-def test_validator_dangling_detection_valid():
-    validator = DanglingDetectionValidator()
-    rule = SigmaRule.from_yaml(
-        """
-    title: Test
-    status: test
-    logsource:
-        category: test
-    detection:
-        referenced1:
-            field1: val1
-        referenced2:
-            field2: val2
-        referenced3:
-            field3: val3
-        condition: (referenced1 or referenced2) and referenced3
-    """
-    )
-    assert validator.validate(rule) == []
-
-
-def test_validator_dangling_detection_valid_x_of_wildcard():
-    validator = DanglingDetectionValidator()
-    rule = SigmaRule.from_yaml(
-        """
-    title: Test
-    status: test
-    logsource:
-        category: test
-    detection:
-        referenced1:
-            field1: val1
-        referenced2:
-            field2: val2
-        referenced3:
-            field3: val3
-        condition: 1 of referenced*
-    """
-    )
-    assert validator.validate(rule) == []
-
-
-def test_validator_dangling_detection_valid_x_of_them():
-    validator = DanglingDetectionValidator()
-    rule = SigmaRule.from_yaml(
-        """
-    title: Test
-    status: test
-    logsource:
-        category: test
-    detection:
-        referenced1:
-            field1: val1
-        referenced2:
-            field2: val2
-        referenced3:
-            field3: val3
-        condition: 1 of them
-    """
-    )
-    assert validator.validate(rule) == []
-
-
-def test_validator_them_condition_with_single_detection():
-    validator = ThemConditionWithSingleDetectionValidator()
-    rule = SigmaRule.from_yaml(
-        """
-    title: Test
-    status: test
-    logsource:
-        category: test
-    detection:
-        selection:
-            field1: val1
-        condition: 1 of them
-    """
-    )
-    assert validator.validate(rule) == [ThemConditionWithSingleDetectionIssue([rule])]
-
-
-def test_validator_them_condition_with_multiple_detection():
-    validator = ThemConditionWithSingleDetectionValidator()
-    rule = SigmaRule.from_yaml(
-        """
-    title: Test
-    status: test
-    logsource:
-        category: test
-    detection:
-        selection1:
-            field1: val1
-        selection2:
-            field2: val2
-        condition: 1 of them
-    """
-    )
-    assert validator.validate(rule) == []
-
-
-def test_validator_all_of_them():
-    validator = AllOfThemConditionValidator()
-    rule = SigmaRule.from_yaml(
-        """
-    title: Test
-    status: test
-    logsource:
-        category: test
-    detection:
-        selection1:
-            field1: val1
-        selection2:
-            field2: val2
-        condition: all of them
-    """
-    )
-    assert validator.validate(rule) == [AllOfThemConditionIssue([rule])]
 
 
 def test_validator_double_wildcard():
@@ -425,174 +265,6 @@ def test_validator_wildcards_instead_of_modifiers_inconsistent():
               - "*val1*"
               - "*val2"
               - "val3*"
-        condition: sel
-    """
-    )
-    assert validator.validate(rule) == []
-
-
-def test_validator_all_without_contains():
-    validator = InvalidModifierCombinationsValidator()
-    rule = SigmaRule.from_yaml(
-        """
-    title: Test
-    status: test
-    logsource:
-        category: test
-    detection:
-        sel:
-            field|all:
-                - value1
-                - value2
-                - value3
-        condition: sel
-    """
-    )
-    assert validator.validate(rule) == [
-        AllWithoutContainsModifierIssue(
-            [rule],
-            SigmaDetectionItem("field", [SigmaAllModifier], ["value1", "value2", "value3"]),
-        )
-    ]
-
-
-def test_validator_all_without_contains_unbound():
-    validator = InvalidModifierCombinationsValidator()
-    rule = SigmaRule.from_yaml(
-        """
-    title: Test
-    status: test
-    logsource:
-        category: test
-    detection:
-        sel:
-            "|all":
-                - value1
-                - value2
-                - value3
-        condition: sel
-    """
-    )
-    assert validator.validate(rule) == []
-
-
-def test_validator_all_with_contains():
-    validator = InvalidModifierCombinationsValidator()
-    rule = SigmaRule.from_yaml(
-        """
-    title: Test
-    status: test
-    logsource:
-        category: test
-    detection:
-        sel:
-            field|contains|all:
-                - value1
-                - value2
-                - value3
-        condition: sel
-    """
-    )
-    assert validator.validate(rule) == []
-
-
-def test_validator_base64offset_without_contains_modifier():
-    validator = InvalidModifierCombinationsValidator()
-    rule = SigmaRule.from_yaml(
-        """
-    title: Test
-    status: test
-    logsource:
-        category: test
-    detection:
-        sel:
-            field|base64offset: value
-        condition: sel
-    """
-    )
-    assert validator.validate(rule) == [
-        Base64OffsetWithoutContainsModifierIssue(
-            [rule], SigmaDetectionItem("field", [SigmaBase64OffsetModifier], ["value"])
-        )
-    ]
-
-
-def test_validator_base64offset_after_contains_modifier():
-    with pytest.raises(SigmaValueError, match="strings with wildcards"):
-        rule = SigmaRule.from_yaml(
-            """
-        title: Test
-        status: test
-        logsource:
-            category: test
-        detection:
-            sel:
-                field|contains|base64offset: value
-            condition: sel
-        """
-        )
-
-
-def test_validator_base64offset_with_contains_modifier():
-    validator = InvalidModifierCombinationsValidator()
-    rule = SigmaRule.from_yaml(
-        """
-    title: Test
-    status: test
-    logsource:
-        category: test
-    detection:
-        sel:
-            field|base64offset|contains: value
-        condition: sel
-    """
-    )
-    assert validator.validate(rule) == []
-
-
-def test_validator_multiple_modifier():
-    validator = InvalidModifierCombinationsValidator()
-    rule = SigmaRule.from_yaml(
-        """
-    title: Test
-    status: test
-    logsource:
-        category: test
-    detection:
-        sel:
-            field|base64offset|base64offset|contains|contains: value
-        condition: sel
-    """
-    )
-    assert validator.validate(rule) == [
-        ModifierAppliedMultipleIssue(
-            [rule],
-            SigmaDetectionItem(
-                "field",
-                [
-                    SigmaBase64OffsetModifier,
-                    SigmaBase64OffsetModifier,
-                    SigmaContainsModifier,
-                    SigmaContainsModifier,
-                ],
-                ["value"],
-            ),
-            {SigmaBase64OffsetModifier, SigmaContainsModifier},
-        )
-    ]
-
-
-def test_validator_multiple_base64_modifier():
-    validator = InvalidModifierCombinationsValidator()
-    rule = SigmaRule.from_yaml(
-        """
-    title: Test
-    status: test
-    logsource:
-        category: test
-    detection:
-        sel:
-            field|base64|base64: value
         condition: sel
     """
     )
